+++ Common/Common.Messaging/Events/AuctionEndingSoonEvent.cs
@@ -4,7 +4,7 @@ public class AuctionEndingSoonEvent
 {
     public Guid AuctionId { get; set; }
     public string Title { get; set; } = string.Empty;
-    public int CurrentHighBid { get; set; }
+    public decimal CurrentHighBid { get; set; }
     public DateTime EndTime { get; set; }
     public string TimeRemaining { get; set; } = string.Empty;
     public List<string> WatcherUsernames { get; set; } = new();
