+++ Common/Common.Messaging/Events/AuctionStartedEvent.cs
@@ -7,5 +7,5 @@ public class AuctionStartedEvent
     public string Title { get; set; } = string.Empty;
     public DateTime StartTime { get; set; }
     public DateTime EndTime { get; set; }
-    public int ReservePrice { get; set; }
+    public decimal ReservePrice { get; set; }
 }
